and so it begins
