+++ lib/sportradar/api/soccer-old/boxscore.rb
