+++ lib/sportradar/api/soccer-old/category.rb
