+++ lib/sportradar/api/soccer-old/hierarchy.rb
