+++ lib/sportradar/api/soccer-old/player.rb
