+++ lib/sportradar/api/soccer-old/ranking.rb
