+++ lib/sportradar/api/soccer-old/schedule.rb
