+++ lib/sportradar/api/soccer-old/season.rb
