+++ lib/sportradar/api/soccer-old/standing.rb
