+++ lib/sportradar/api/soccer-old/summary.rb
