+++ lib/sportradar/api/soccer-old/team.rb
