+++ lib/sportradar/api/soccer-old/tournament.rb
