+++ lib/sportradar/api/soccer-old/tournament_group.rb
