+++ soccer-old/player_test.rb
