+++ soccer-old/standing_test.rb
